+++ app/app.go
@@ -1,7 +1,9 @@
 package app
 
 import (
-	"github.com/go-redis/redis"
+	"context"
+
+	"github.com/go-redis/redis/v8"
 	"github.com/jmoiron/sqlx"
 	"github.com/keratin/authn-server/app/data"
 	"github.com/keratin/authn-server/lib/oauth"
@@ -36,9 +38,9 @@ func NewApp(cfg *Config, logger logrus.FieldLogger) (*App, error) {
 		return nil, errors.Wrap(err, "data.NewDB")
 	}
 
-	var redis *redis.Client
+	var redis redis.UniversalClient
 	if cfg.RedisURL != nil {
-		redis, err = dataRedis.New(cfg.RedisURL)
+		redis, err = dataRedis.New(cfg.RedisURL, cfg.RedisCluster, cfg.RedisMasterName)
 		if err != nil {
 			return nil, errors.Wrap(err, "redis.New")
 		}
@@ -106,7 +108,7 @@ func NewApp(cfg *Config, logger logrus.FieldLogger) (*App, error) {
 		// Provide access to root DB - useful when extending AccountStore functionality
 		DB:                db,
 		DbCheck:           func() bool { return db.Ping() == nil },
-		RedisCheck:        func() bool { return redis != nil && redis.Ping().Err() == nil },
+		RedisCheck:        func() bool { return redis != nil && redis.Ping(context.Background()).Err() == nil },
 		Config:            cfg,
 		AccountStore:      accountStore,
 		RefreshTokenStore: tokenStore,