+++ app/data/redis/refresh_token_store.go
@@ -1,19 +1,20 @@
 package redis
 
 import (
+	"context"
 	"encoding/hex"
 	"fmt"
 	"strconv"
 	"time"
 
-	"github.com/go-redis/redis"
-	"github.com/keratin/authn-server/lib"
+	"github.com/go-redis/redis/v8"
 	"github.com/keratin/authn-server/app/models"
+	"github.com/keratin/authn-server/lib"
 )
 
 type RefreshTokenStore struct {
-	*redis.Client
-	TTL time.Duration
+	Client redis.UniversalClient
+	TTL    time.Duration
 }
 
 // Redis key for token => accountID lookup
@@ -33,7 +34,7 @@ func (s *RefreshTokenStore) Find(hexToken models.RefreshToken) (int, error) {
 	if err != nil {
 		return 0, err
 	}
-	str, err := s.Client.Get(keyForToken(binToken)).Result()
+	str, err := s.Client.Get(context.Background(), keyForToken(binToken)).Result()
 	if err == redis.Nil {
 		return 0, nil
 	} else if err != nil {
@@ -48,16 +49,17 @@ func (s *RefreshTokenStore) Touch(hexToken models.RefreshToken, accountID int) e
 		return err
 	}
 
-	_, err = s.Client.Pipelined(func(pipe redis.Pipeliner) error {
-		pipe.Expire(keyForToken(binToken), s.TTL)
-		pipe.Expire(keyForAccount(accountID), s.TTL)
+	ctx := context.Background()
+	_, err = s.Client.Pipelined(ctx, func(pipe redis.Pipeliner) error {
+		pipe.Expire(ctx, keyForToken(binToken), s.TTL)
+		pipe.Expire(ctx, keyForAccount(accountID), s.TTL)
 		return nil
 	})
 	return err
 }
 
 func (s *RefreshTokenStore) FindAll(accountID int) ([]models.RefreshToken, error) {
-	bins, err := s.Client.SMembers(keyForAccount(accountID)).Result()
+	bins, err := s.Client.SMembers(context.Background(), keyForAccount(accountID)).Result()
 	if err != nil {
 		return nil, err
 	}
@@ -76,13 +78,14 @@ func (s *RefreshTokenStore) Create(accountID int) (models.RefreshToken, error) {
 		return "", err
 	}
 
-	_, err = s.Client.Pipelined(func(pipe redis.Pipeliner) error {
+	ctx := context.Background()
+	_, err = s.Client.Pipelined(ctx, func(pipe redis.Pipeliner) error {
 		// persist the token
-		pipe.Set(keyForToken(binToken), accountID, s.TTL)
+		pipe.Set(ctx, keyForToken(binToken), accountID, s.TTL)
 
 		// maintain a list of tokens per accountID
-		pipe.SAdd(keyForAccount(accountID), binToken)
-		pipe.Expire(keyForAccount(accountID), s.TTL)
+		pipe.SAdd(ctx, keyForAccount(accountID), binToken)
+		pipe.Expire(ctx, keyForAccount(accountID), s.TTL)
 
 		return nil
 	})
@@ -102,14 +105,15 @@ func (s *RefreshTokenStore) Revoke(hexToken models.RefreshToken) error {
 		return nil
 	}
 
-	_, err = s.Client.Pipelined(func(pipe redis.Pipeliner) error {
+	ctx := context.Background()
+	_, err = s.Client.Pipelined(ctx, func(pipe redis.Pipeliner) error {
 		binToken, err := hex.DecodeString(string(hexToken))
 		if err != nil {
 			return err
 		}
 
-		pipe.Del(keyForToken(binToken))
-		pipe.SRem(keyForAccount(accountID), binToken)
+		pipe.Del(ctx, keyForToken(binToken))
+		pipe.SRem(ctx, keyForAccount(accountID), binToken)
 
 		return nil
 	})